Cleanup and more documentation.
